Fix hard coded formatting strings for 64 bit numbers (#1831)
Postgres provides OS agnosting formatting macros for
formatting 64 bit numbers. Replaced %ld %lu with
INT64_FORMAT and UINT64_FORMAT respectively.

Also found some incorrect usages of formatting
flags and fixed them.

+++ src/backend/distributed/executor/multi_client_executor.c
@@ -937,8 +937,8 @@ MultiClientWait(WaitInfo *waitInfo)
 		else if (rc == 0)
 		{
 			ereport(DEBUG5,
-					(errmsg("waiting for activity on tasks took longer than %ld ms",
-							(long) RemoteTaskCheckInterval * 10)));
+					(errmsg("waiting for activity on tasks took longer than %d ms",
+							(int) RemoteTaskCheckInterval * 10)));
 		}
 
 		/*
